+++ src/commands/transaction_commands.rs
@@ -33,7 +33,7 @@ pub trait TransactionCommands<'a>: Sized {
     /// [<https://redis.io/commands/watch/>](https://redis.io/commands/watch/)
     #[must_use]
     fn watch(self, keys: impl Serialize) -> PreparedCommand<'a, Self, ()> {
-        prepare_command(self, cmd("WATCH").key(keys))
+        prepare_command(self, cmd("WATCH").keys(keys))
     }
 
     /// Flushes all the previously watched keys for a transaction.